Fix integer literal typing in Haunted-C
- Parser: unsuffixed integer literals get SpecInt type (mathematical
  integer) instead of hardcoded unsigned int32. Suffixed literals (u, l,
  ll, z) get proper C types based on suffix and target widths.
- Emitter: add SpecInt IntLit emission (bare number) and SpecInt→Int
  cast emission (int_to_t / uint_to_t).
- Elaborator: fix meet_type for Int/Int with different signedness/width
  using C-style usual arithmetic conversions instead of returning None.
- Tests: remove 12 unnecessary (int) and (_specint) casts that are no
  longer needed.

+++ test/antiquot.c
@@ -8,8 +8,8 @@ void test_rvalue_antiquot(int *x)
 }
 
 void test_ghost_stmt(int *x)
-    _requires(*x == (int) 0)
-    _ensures(*x == (int) 0)
+    _requires(*x == 0)
+    _ensures(*x == 0)
 {
     _ghost_stmt(rewrite (pure ($(*x) = $(*x))) as (pure ($(*x) = $(*x))));
     return;
